Got rid of annoying overflow issue
